requirements.txt
streamlit
PyMuPDF
python-docx
numpy
faiss-cpu
requests
PyYAML
tqdm
sentencepiece
